# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Consultas Médicas",
        "description": "Documentação da API do Sistema de Consultas Médicas.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/tipos-consulta": {
            "get": {
                "tags": [
                    "Tipos de Consulta"
                ],
                "summary": "Lista todos os tipos de consulta",
                "description": "Retorna todos os tipos de consulta cadastrados.",
                "operationId": "1978966154aab67c788ab7ee63c44962",
                "responses": {
                    "200": {
                        "description": "Lista de tipos de consulta retornada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "data": [
                                        {
                                            "id": 1,
                                            "nome": "Consulta Cardiológica",
                                            "duracao": 30,
                                            "hora_inicio": "08:00:00",
                                            "hora_fim": "18:00:00",
                                            "created_at": "2026-07-22T23:07:45.000000Z",
                                            "updated_at": "2026-07-22T23:07:45.000000Z"
                                        },
                                        {
                                            "id": 2,
                                            "nome": "Consulta Ortopédica",
                                            "duracao": 45,
                                            "hora_inicio": "09:00:00",
                                            "hora_fim": "17:00:00",
                                            "created_at": "2026-07-22T23:07:45.000000Z",
                                            "updated_at": "2026-07-22T23:07:45.000000Z"
                                        },
                                        {
                                            "id": 3,
                                            "nome": "Consulta Dermatológica",
                                            "duracao": 30,
                                            "hora_inicio": "10:00:00",
                                            "hora_fim": "16:00:00",
                                            "created_at": "2026-07-22T23:07:45.000000Z",
                                            "updated_at": "2026-07-22T23:07:45.000000Z"
                                        },
                                        {
                                            "id": 4,
                                            "nome": "Consulta Neurológica",
                                            "duracao": 50,
                                            "hora_inicio": "08:00:00",
                                            "hora_fim": "15:30:00",
                                            "created_at": "2026-07-22T23:07:45.000000Z",
                                            "updated_at": "2026-07-22T23:19:39.000000Z"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tipos de Consulta"
                ],
                "summary": "Cadastra um novo tipo de consulta",
                "description": "Cria um novo tipo de consulta no sistema.",
                "operationId": "8890bc7736019740d9b5e34e3a1cbec5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "duracao",
                                    "especialidades",
                                    "hora_inicio",
                                    "hora_fim"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Consulta Cardiológica"
                                    },
                                    "duracao": {
                                        "type": "integer",
                                        "example": 60
                                    },
                                    "especialidades": {
                                        "description": "IDs das especialidades separados por vírgula",
                                        "type": "string",
                                        "example": "1,3"
                                    },
                                    "hora_inicio": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "08:00"
                                    },
                                    "hora_fim": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "18:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tipo de consulta cadastrado com sucesso."
                    },
                    "422": {
                        "description": "Erro de validação."
                    }
                }
            }
        },
        "/api/tipos-consulta/{id}": {
            "get": {
                "tags": [
                    "Tipos de Consulta"
                ],
                "summary": "Busca um tipo de consulta pelo ID",
                "description": "Retorna um tipo de consulta específico.",
                "operationId": "f1308813a10afa2053736a06d55d223b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do tipo de consulta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipo de consulta encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "data": {
                                        "id": 1,
                                        "nome": "Consulta Cardiológica",
                                        "duracao": 30,
                                        "hora_inicio": "08:00:00",
                                        "hora_fim": "18:00:00",
                                        "created_at": "2026-07-22T23:07:45.000000Z",
                                        "updated_at": "2026-07-22T23:07:45.000000Z"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tipo de consulta não encontrado."
                    }
                }
            },
            "put": {
                "tags": [
                    "Tipos de Consulta"
                ],
                "summary": "Atualiza um tipo de consulta",
                "description": "Atualiza os dados de um tipo de consulta existente.",
                "operationId": "ad728e34fb72b5839ddd11cb24f0053f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do tipo de consulta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Consulta Neurológica"
                                    },
                                    "especialidade_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tipo de consulta atualizado com sucesso."
                    },
                    "404": {
                        "description": "Tipo de consulta não encontrado."
                    },
                    "422": {
                        "description": "Erro de validação."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tipos de Consulta"
                ],
                "summary": "Remove um tipo de consulta",
                "description": "Exclui um tipo de consulta pelo ID.",
                "operationId": "ab1a4c4ee00e1d6ed781934e8ac50f23",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do tipo de consulta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipo de consulta removido com sucesso."
                    },
                    "404": {
                        "description": "Tipo de consulta não encontrado."
                    }
                }
            }
        }
    },
...
}